test: organize e2e tests into logical subdirectories
Refactored the `e2e/tests/` structure to group tests by functionality
(auth, config, error, observability, proxy, response, security) for
better maintainability. Updated `Cargo.toml` to reflect the new paths.
