+++ e2e/tests/auth/mod.rs
@@ -1,0 +1,5 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod basic;
+mod forward;
