+++ e2e/tests/config/mod.rs
@@ -1,0 +1,5 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod directives;
+mod include;
