+++ e2e/tests/error/mod.rs
@@ -1,0 +1,6 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod handle;
+mod intercept;
+mod page;
