+++ e2e/tests/observability/mod.rs
@@ -1,0 +1,6 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod admin;
+mod metrics;
+mod traces;
