+++ e2e/tests/observability/traces.rs
@@ -7,14 +7,13 @@ use testcontainers::{
     runners::AsyncRunner,
 };
 
-mod common;
 
 async fn create_ferron_container(
     network: &str,
     webroot_dir: &Path,
     config_file: &Path,
 ) -> Result<ContainerAsync<GenericImage>, TestcontainersError> {
-    let ferron_image = self::common::build_ferron_image().await?;
+    let ferron_image = crate::common::build_ferron_image().await?;
     ferron_image
         .with_exposed_port(ContainerPort::Tcp(80))
         .with_wait_for(WaitFor::Http(Box::new(
@@ -43,7 +42,7 @@ async fn create_otlp_container(
     let mut attempts = 0;
     loop {
         attempts += 1;
-        let otlp_image = self::common::build_otlp_image().await?;
+        let otlp_image = crate::common::build_otlp_image().await?;
         let start_res = otlp_image
             .with_exposed_port(ContainerPort::Tcp(4318))
             // short fixed wait; test will poll the mock collector endpoint for received payloads
@@ -77,9 +76,9 @@ async fn test_otlp_traces_exported() {
     nix::sys::stat::umask(nix::sys::stat::Mode::from_bits(0o000).unwrap());
 
     #[cfg(unix)]
-    let webroot_dir = common::create_temp_dir();
+    let webroot_dir = crate::common::create_temp_dir();
     #[cfg(unix)]
-    let mut config_file = common::create_temp_file();
+    let mut config_file = crate::common::create_temp_file();
 
     #[cfg(not(unix))]
     let webroot_dir = tempfile::tempdir().unwrap();
@@ -104,7 +103,7 @@ async fn test_otlp_traces_exported() {
         )
         .unwrap();
 
-    common::write_file(webroot_dir.path().join("basic.txt"), b"hello").unwrap();
+    crate::common::write_file(webroot_dir.path().join("basic.txt"), b"hello").unwrap();
 
     let network = "e2e-test-otlp";
