+++ e2e/tests/proxy/health.rs
@@ -7,12 +7,11 @@ use testcontainers::{
     runners::AsyncRunner,
 };
 
-mod common;
 
 async fn create_backend_container(
     network: &str,
 ) -> Result<ContainerAsync<GenericImage>, TestcontainersError> {
-    let backend_image = self::common::build_backend_image().await?;
+    let backend_image = crate::common::build_backend_image().await?;
     backend_image
         .with_exposed_port(ContainerPort::Tcp(3000))
         .with_wait_for(WaitFor::Http(Box::new(
@@ -31,7 +30,7 @@ async fn create_ferron_container(
     network: &str,
     config_file: &std::path::Path,
 ) -> Result<ContainerAsync<GenericImage>, TestcontainersError> {
-    let ferron_image = self::common::build_ferron_image().await?;
+    let ferron_image = crate::common::build_ferron_image().await?;
     ferron_image
         .with_exposed_port(ContainerPort::Tcp(80))
         .with_wait_for(WaitFor::Http(Box::new(
@@ -66,7 +65,7 @@ impl HealthCheckTestContext {
         nix::sys::stat::umask(nix::sys::stat::Mode::from_bits(0o000).unwrap());
 
         #[cfg(unix)]
-        let mut config_file = common::create_temp_file();
+        let mut config_file = crate::common::create_temp_file();
         #[cfg(not(unix))]
         let mut config_file = tempfile::NamedTempFile::new().unwrap();
