+++ e2e/tests/proxy/mod.rs
@@ -1,0 +1,10 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod affinity;
+mod grpc;
+mod health;
+mod lb;
+mod proxy_header;
+mod rproxy;
+mod srv;
