+++ e2e/tests/response/mod.rs
@@ -1,0 +1,7 @@
+#[path = "../common/mod.rs"]
+mod common;
+
+mod headers;
+mod redirect;
+mod replace;
+mod rewrite;
