+++ e2e/tests/security/abuseban.rs
@@ -1,6 +1,6 @@
 use testcontainers::core::ContainerPort;
 
-mod common;
+use crate::common;
 
 #[tokio::test]
 async fn test_abuse_protection_does_not_block_normal_traffic() {
@@ -33,7 +33,7 @@ async fn test_abuse_protection_does_not_block_normal_traffic() {
     )
     .unwrap();
 
-    let container = common::create_ferron_container(webroot_dir.path(), config_file.path())
+    let container = crate::create_ferron_container(webroot_dir.path(), config_file.path())
         .await
         .unwrap();
 
@@ -43,7 +43,6 @@ async fn test_abuse_protection_does_not_block_normal_traffic() {
         .unwrap();
     let client = reqwest::Client::builder().no_proxy().build().unwrap();
 
-    // Requests should pass through since no events have been recorded to trigger a ban
     let response = client
         .get(format!("http://localhost:{}/test.txt", port))
         .send()
@@ -58,7 +57,6 @@ async fn test_abuse_protection_does_not_block_normal_traffic() {
         .unwrap();
     assert_eq!(response.status(), 200);
 
-    // Verify the server properly handles concurrent requests with abuse protection
     let mut handles = Vec::new();
     for _ in 0..10 {
         let url = format!("http://localhost:{}/test.txt", port);
@@ -80,12 +78,6 @@ async fn test_abuse_protection_blocks_rate_limit_abusers() {
     let webroot_dir = common::create_temp_dir();
     let config_file = common::create_temp_file();
 
-    // Configure rate_limit to allow only 1 request, and abuse_protection
-    // to ban after 2 rate limit events. Sending 3 fast requests should:
-    //   1st → pass (consumes the only token)
-    //   2nd → 429 (rate limited, emits abuse event #1)
-    //   3rd → 429 (rate limited, emits abuse event #2 → ban triggered)
-    //   4th → 403 (banned by abuse protection)
     common::write_file(
         config_file.path().to_path_buf(),
         r#"
@@ -113,7 +105,7 @@ async fn test_abuse_protection_blocks_rate_limit_abusers() {
 
     common::write_file(webroot_dir.path().join("data.txt").to_path_buf(), b"data").unwrap();
 
-    let container = common::create_ferron_container(webroot_dir.path(), config_file.path())
+    let container = crate::create_ferron_container(webroot_dir.path(), config_file.path())
         .await
         .unwrap();
 
@@ -124,19 +116,15 @@ async fn test_abuse_protection_blocks_rate_limit_abusers() {
     let client = reqwest::Client::builder().no_proxy().build().unwrap();
     let url = format!("http://localhost:{}/data.txt", port);
 
-    // 1st request: passes (consumes the token)
     let resp = client.get(&url).send().await.unwrap();
     assert_eq!(resp.status(), 200, "first request should pass");
 
-    // 2nd request: rate limited (429), triggers abuse event #1
     let resp = client.get(&url).send().await.unwrap();
     assert_eq!(resp.status(), 429, "second request should be rate limited");
 
-    // 3rd request: rate limited (429), triggers abuse event #2 → ban
     let resp = client.get(&url).send().await.unwrap();
     assert_eq!(resp.status(), 429, "third request should be rate limited");
 
-    // 4th request: IP is now banned → 403 Forbidden with Retry-After
     let resp = client.get(&url).send().await.unwrap();
     assert_eq!(
         resp.status(),
@@ -144,7 +132,6 @@ async fn test_abuse_protection_blocks_rate_limit_abusers() {
         "fourth request should be banned by abuse protection"
     );
 
-    // Verify the Retry-After header is present
     let retry_after = resp.headers().get("retry-after");
     assert!(
         retry_after.is_some(),
@@ -161,8 +148,6 @@ async fn test_abuse_protection_without_rate_limit_does_not_block() {
     let webroot_dir = common::create_temp_dir();
     let config_file = common::create_temp_file();
 
-    // Abuse protection without rate_limit — no events are emitted so
-    // no bans should be triggered regardless of request volume.
     common::write_file(
         config_file.path().to_path_buf(),
         r#"
@@ -184,7 +169,7 @@ async fn test_abuse_protection_without_rate_limit_does_not_block() {
 
     common::write_file(webroot_dir.path().join("page.html").to_path_buf(), b"page").unwrap();
 
-    let container = common::create_ferron_container(webroot_dir.path(), config_file.path())
+    let container = crate::create_ferron_container(webroot_dir.path(), config_file.path())
         .await
         .unwrap();
 
@@ -194,7 +179,6 @@ async fn test_abuse_protection_without_rate_limit_does_not_block() {
         .unwrap();
     let client = reqwest::Client::builder().no_proxy().build().unwrap();
 
-    // Many requests with no rate limiting should all pass
     for _ in 0..20 {
         let resp = client
             .get(format!("http://localhost:{}/page.html", port))
