+++ e2e/tests/security/limit.rs
@@ -1,5 +1,7 @@
 use std::io::Write;
 
+use testcontainers::core::ContainerPort;
+
 use crate::{common, create_ferron_container};
 
 /// Test for rate limiting race condition fix.
@@ -11,7 +13,6 @@ async fn test_rate_limiting_race_condition_fixed() {
     let webroot_dir = common::create_temp_dir();
     let mut config_file = common::create_temp_file();
 
-    // Simple rate limiting test with low limit to trigger quickly
     config_file
         .as_file_mut()
         .write_all(
@@ -43,7 +44,6 @@ async fn test_rate_limiting_race_condition_fixed() {
             .expect("Failed to get host port")
     );
 
-    // Send multiple requests sequentially to stay within bucket window
     let mut allowed = 0;
     let mut rejected = 0;
     for _ in 0..10 {
@@ -59,7 +59,6 @@ async fn test_rate_limiting_race_condition_fixed() {
         }
     }
 
-    // With 5r/s limit, we expect ~5 allowed, ~5 rejected
     println!(
         "Rate limiting test: allowed={}, rejected={} (expected ~5 each)",
         allowed, rejected
@@ -70,4 +69,65 @@ async fn test_rate_limiting_race_condition_fixed() {
         rejected,
         allowed
     );
+}
+
+/// Basic rate limiting: rate 2 / burst 2 should trigger 429 under load.
+#[tokio::test]
+async fn test_rate_limiting_basic() {
+    let _ = rustls::crypto::ring::default_provider().install_default();
+
+    let webroot_dir = common::create_temp_dir();
+    let mut config_file = common::create_temp_file();
+
+    config_file
+        .as_file_mut()
+        .write_all(
+            r#"
+*:80 {
+  root "/var/www/ferron"
+  rate_limit {
+    rate 2
+    burst 2
+  }
+}
+"#
+            .as_bytes(),
+        )
+        .unwrap();
+
+    common::write_file(webroot_dir.path().join("test.txt"), b"test content").unwrap();
+    common::write_file(webroot_dir.path().join("basic.txt"), b"basic content").unwrap();
+
+    let container = create_ferron_container(webroot_dir.path(), config_file.path())
+        .await
+        .unwrap();
+
+    let port = container
+        .get_host_port_ipv4(ContainerPort::Tcp(80))
+        .await
+        .unwrap();
+    let client = reqwest::Client::new();
+
+    for _ in 0..5 {
+        client
+            .get(format!("http://localhost:{}/test.txt", port))
+            .send()
+            .await
+            .ok();
+        client
+            .get(format!("http://localhost:{}/test.txt", port))
+            .send()
+            .await
+            .ok();
+    }
+
+    let response = client
+        .get(format!("http://localhost:{}/basic.txt", port))
+        .send()
+        .await
+        .unwrap();
+
+    assert_eq!(response.status(), reqwest::StatusCode::TOO_MANY_REQUESTS);
+
+    container.stop().await.unwrap();
 }
